LayoutTests/http/tests/security/xss-DENIED-method-with-iframe-proto-expected.txt

CONSOLE MESSAGE: line 1: Unsafe JavaScript attempt to access frame with URL http://localhost:8000/ from frame with URL http://127.0.0.1:8000/security/xss-DENIED-method-with-iframe-proto.html. Domains, protocols and ports must match.

CONSOLE MESSAGE: line 1: Unsafe JavaScript attempt to access frame with URL http://localhost:8000/ from frame with URL http://127.0.0.1:8000/security/xss-DENIED-method-with-iframe-proto.html. Domains, protocols and ports must match.

CONSOLE MESSAGE: line 1: Unsafe JavaScript attempt to access frame with URL http://localhost:8000/ from frame with URL http://127.0.0.1:8000/security/xss-DENIED-method-with-iframe-proto.html. Domains, protocols and ports must match.


Tests that making other frame window a prototype doesn't expose that window methods
PASS: this.wasInvoked should be 'false' and is.
